apps/azure-rag-setup/requirements.txt
# Azure RAG Setup Dependencies
# Install with: pip install -r requirements.txt

# Core Azure SDK packages
azure-storage-blob==12.19.0
azure-search-documents==11.4.0
azure-identity==1.15.0
azure-core==1.29.5

# Environment and configuration
python-dotenv==1.0.0

# HTTP requests and utilities
requests==2.31.0
urllib3==2.0.7

# Data processing and utilities
pathlib2==2.3.7.post1
json5==0.9.14

# Retry and progress tracking
tenacity==8.2.3
tqdm==4.66.1

# Microsoft Graph API dependencies
msal==1.26.0                  # Microsoft Authentication Library
msgraph-core==1.0.0           # Microsoft Graph SDK
pyyaml==6.0.1                 # Config file parsing

# Office 365 SharePoint integration
office365-rest-python-client==2.5.0

# Testing and validation
pytest==7.4.3
pytest-cov==4.1.0

# Enhanced functionality (optional)
# azure-keyvault-secrets==4.7.0  # For production secret management
# azure-monitor-opentelemetry==1.0.0  # For monitoring
